+++ ArduinoBaseControl/dynamics.hpp
@@ -7,9 +7,9 @@
  */
 class dynamics_t {
    private:
-    double inertia_mass_;            // 慣性モーメント (kg*m^2)
-    double viscous_friction_coeff_;  // 粘性摩擦係数 (N*m*s/rad)
-    double torque_constant_;         // モーターのトルク定数 (Nm/A)
+    float inertia_mass_;            // 慣性モーメント (kg*m^2)
+    float viscous_friction_coeff_;  // 粘性摩擦係数 (N*m*s/rad)
+    float torque_constant_;         // モーターのトルク定数 (Nm/A)
 
    public:
     /**
@@ -18,7 +18,7 @@ class dynamics_t {
      * @param viscous_friction_coeff 粘性摩擦係数
      * @param torque_constant モーターのトルク定数
      */
-    dynamics_t(double inertia_mass, double viscous_friction_coeff, double torque_constant);
+    dynamics_t(float inertia_mass, float viscous_friction_coeff, float torque_constant);
 
     /**
      * 動力学計算に基づき、フィードフォワード制御量（トルク）を計算する関数。
@@ -28,22 +28,22 @@ class dynamics_t {
      * @param target_accel 目標加速度 (単位: rad/s^2)
      * @return フィードフォワード制御量（トルク）
      */
-    double calculate_feedforward_control(double target_vel, double target_accel) const;
+    float calculate_feedforward_control(float target_vel, float target_accel) const;
 
     /**
      * フィードフォワード制御量を電流値に変換する関数
      * @param feedforward_torque_force フィードフォワード制御量（トルク）
      * @return 電流値 (A)
      */
-    double convert_to_current_command(double feedforward_torque_force) const;
+    float convert_to_current_command(float feedforward_torque_force) const;
 
     // ゲッター関数
-    double get_inertia_mass() const { return inertia_mass_; }
-    double get_viscous_friction_coeff() const { return viscous_friction_coeff_; }
-    double get_torque_constant() const { return torque_constant_; }
+    float get_inertia_mass() const { return inertia_mass_; }
+    float get_viscous_friction_coeff() const { return viscous_friction_coeff_; }
+    float get_torque_constant() const { return torque_constant_; }
 
     // セッター関数
-    void set_inertia_mass(double inertia_mass) { inertia_mass_ = inertia_mass; }
-    void set_viscous_friction_coeff(double viscous_friction_coeff) { viscous_friction_coeff_ = viscous_friction_coeff; }
-    void set_torque_constant(double torque_constant) { torque_constant_ = torque_constant; }
+    void set_inertia_mass(float inertia_mass) { inertia_mass_ = inertia_mass; }
+    void set_viscous_friction_coeff(float viscous_friction_coeff) { viscous_friction_coeff_ = viscous_friction_coeff; }
+    void set_torque_constant(float torque_constant) { torque_constant_ = torque_constant; }
 };
